requirements.txt
asgiref==3.2.7
autopep8==1.5.3
Django==3.0.7
django-crispy-forms==1.9.1
django-widget-tweaks==1.4.8
gunicorn==20.0.4
Pillow==7.1.2
pycodestyle==2.6.0
pytz==2020.1
sqlparse==0.3.1
toml==0.10.1
